Homework 2: Work_history Longer branch commit

+++ src/Git_and_java_basics/Work_history/src/main/java/org/example/Main.java
@@ -3,7 +3,7 @@ package org.example;
 public class Main {
     public static void main(String[] args) {
 
-        SalesManager salesManager = new SalesManager(new int[]{100, 10, 20, 300, 15});
+        SalesManager salesManager = new SalesManager(new long[]{100, 10, 20, 300, 15});
 
         System.out.println(salesManager.max());
     }
